+++ modroc_controller/sensors/include/bme280.h
@@ -38,4 +38,4 @@ Error_Returns bme280_init(uint32_t id, uint32_t address);
 
 Error_Returns bme280_reset(uint32_t id);
 
-Error_Returns bme280_get_current_pressure(uint32_t id, double *pressure_ptr);
+Error_Returns bme280_get_current_pressure(uint32_t id, uint32_t *pressure_ptr);
